--- src/CSE-Labs/stylecop.json
@@ -1,14 +1,0 @@
-﻿{
-  "$schema": "https://raw.githubusercontent.com/DotNetAnalyzers/StyleCopAnalyzers/master/StyleCop.Analyzers/StyleCop.Analyzers/Settings/stylecop.schema.json",
-  "settings": {
-    "orderingRules": {
-      "usingDirectivesPlacement": "outsideNamespace"
-    },
-
-    "documentationRules": {
-      "xmlHeader": false,
-      "companyName": "Microsoft",
-      "copyrightText": "Copyright (c) Microsoft Corporation. All rights reserved.\nLicensed under the MIT License. See LICENSE in the project root for license information."
-    }
-  }
-}
