+++ src/MC/Controllers/ErrorResult.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System.Net;
+
+namespace MC.Middleware
+{
+    public class ErrorResult
+    {
+        public int Status => (int)Error;
+
+        public string Message { get; set; }
+
+        public HttpStatusCode Error { get; set; }
+    }
+}
