+++ src/MC/Database.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace CSApp.Model
+{
+    /// <summary>
+    /// Database class (singleton)
+    /// </summary>
+    public sealed partial class Database
+    {
+        // placeholder
+    }
+}
