+++ src/appsettings.Development.json
@@ -1,10 +1,8 @@
-﻿{
+{
   "Logging": {
     "LogLevel": {
-      "Default": "Error",
-      "System": "Error",
-      "Microsoft": "Error"
+      "Default": "Warning",
+      "Microsoft.AspNetCore": "Warning"
     }
   }
 }
-
