+++ src/appsettings.json
@@ -1,9 +1,9 @@
-﻿{
+{
   "Logging": {
     "LogLevel": {
-      "Default": "Error",
-      "System": "Error",
-      "Microsoft": "Error"
+      "Default": "Warning",
+      "Microsoft.AspNetCore": "Warning"
     }
-  }
+  },
+  "AllowedHosts": "*"
 }
