+++ src/ascii-art.txt
@@ -1,4 +1,5 @@
-   ___ ___ ___   _         _
-  / __/ __| __| | |   __ _| |__ ___
- | (__\__ \ _|  | |__/ _` | '_ (_-<
-  \___|___/___| |____\__,_|_.__/__/
+   ___ ___     _             
+  / __/ __|   /_\  _ __ _ __ 
+ | (__\__ \  / _ \| '_ \ '_ \
+  \___|___/ /_/ \_\ .__/ .__/
+                  |_|  |_|
