+++ src/cadl/build
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+set -e
+
+. "$NVM_DIR/nvm.sh"
+
+cadl install
+
+mkdir -p ../wwwroot/v1
+
+cadl compile . --output-dir ../wwwroot/v1
