+++ src/cadl/cadl-project.yaml
@@ -1,0 +1,2 @@
+emitters:
+  '@cadl-lang/openapi3': true
