+++ src/cadl/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "CSApp",
+  "dependencies": {
+    "@cadl-lang/compiler": "latest",
+    "@cadl-lang/rest": "latest",
+    "@cadl-lang/openapi3": "latest"
+  },
+  "private": true
+}
