--- src/secrets/Database
@@ -1,1 +1,0 @@
-imdb
