--- src/secrets/Environment
@@ -1,1 +1,0 @@
-dev
