--- src/secrets/Table
@@ -1,1 +1,0 @@
-movies
